refactor - changed bodyWeight in InfantVisit.java to Double data type from Long and as well modified the setters, getters, request and response where it's used. Added changeset to modify the database as well.

+++ src/main/java/org/lamisplus/modules/pmtct/domain/dto/InfantVisitRequestDto.java
@@ -17,7 +17,7 @@ public class InfantVisitRequestDto implements Serializable
     private LocalDate visitDate;
     private String infantHospitalNumber;
     private String ancNumber;
-    private Long bodyWeight;
+    private Double bodyWeight;
     private String visitStatus;
     //private String ctxStatus;
     private String breastFeeding ;
